improve test suite with rust-lang examples

+++ src/heap.rs
@@ -65,7 +65,7 @@ where
 fn heapsort_test() {
     use crate::utils::test_suite;
 
-    test_suite(|arr: &mut [u64], cmp| {
+    test_suite(|arr: &mut [i32], cmp| {
         heapsort(arr, &cmp);
         arr.to_vec()
     });
