+++ src/mergesort.rs
@@ -47,5 +47,5 @@ where
 fn mergesort_test() {
     use crate::utils::test_suite;
 
-    test_suite(|arr: &mut [u64], cmp| mergesort(arr, &cmp));
+    test_suite(|arr: &mut [i32], cmp| mergesort(arr, &cmp));
 }
