+++ src/qsort.rs
@@ -80,12 +80,12 @@ where
 fn quicksort_test() {
     use crate::utils::test_suite;
 
-    test_suite(|arr: &mut [u64], cmp| {
+    test_suite(|arr: &mut [i32], cmp| {
         quicksort(arr, &cmp, PartitionScheme::Lomuto);
         arr.to_vec()
     });
 
-    test_suite(|arr: &mut [u64], cmp| {
+    test_suite(|arr: &mut [i32], cmp| {
         quicksort(arr, &cmp, PartitionScheme::Hoare);
         arr.to_vec()
     });
