+++ src/simple.rs
@@ -50,7 +50,7 @@ where
 fn bubblesort_test() {
     use crate::utils::test_suite;
 
-    test_suite(|arr: &mut [u64], cmp| {
+    test_suite(|arr: &mut [i32], cmp| {
         bubble_sort(arr, &cmp);
         arr.to_vec()
     });
@@ -60,7 +60,7 @@ fn bubblesort_test() {
 fn insertionsort_test() {
     use crate::utils::test_suite;
 
-    test_suite(|arr: &mut [u64], cmp| {
+    test_suite(|arr: &mut [i32], cmp| {
         insertion_sort(arr, &cmp);
         arr.to_vec()
     });
@@ -70,7 +70,7 @@ fn insertionsort_test() {
 fn selectionsort_test() {
     use crate::utils::test_suite;
 
-    test_suite(|arr: &mut [u64], cmp| {
+    test_suite(|arr: &mut [i32], cmp| {
         selection_sort(arr, &cmp);
         arr.to_vec()
     });
